app/requirements.txt
tensorflow==2.19.0
streamlit==1.43.2
numpy==2.1.3
pillow==11.1.0